+++ nifty/cl/operators/outer_product_operator.py
@@ -12,6 +12,7 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
 # Copyright(C) 2013-2020 Max-Planck-Society
+# Copyright(C) 2026 Philipp Arras
 #
 # NIFTy is being developed at the Max-Planck-Institut fuer Astrophysik.
 
@@ -19,6 +20,7 @@ import numpy as np
 
 from ..domain_tuple import DomainTuple
 from ..field import Field
+from ..utilities import iscomplextype
 from .linear_operator import LinearOperator
 
 
@@ -29,13 +31,22 @@ class OuterProduct(LinearOperator):
     ---------
     domain : DomainTuple, the domain of the input field
     field : :class:`nifty.cl.field.Field`
+    flip : bool
+        If False, `field` becomes the leading and the input field the trailing
+        factor of the outer product, i.e. `x -> field * x`. If True, the
+        order is reversed, i.e. `x -> x * field`. Default: False.
     ---------
     """
-    def __init__(self, domain, field):
+    def __init__(self, domain, field, flip=False):
         self._domain = DomainTuple.make(domain)
         self._field = field
-        self._target = DomainTuple.make(
-            tuple(sub_d for sub_d in field.domain._dom + self._domain._dom))
+        self._flip = bool(flip)
+        self._complex = iscomplextype(field.dtype)
+        if self._flip:
+            doms = tuple(self._domain) + tuple(field.domain)
+        else:
+            doms = tuple(field.domain) + tuple(self._domain)
+        self._target = DomainTuple.make(doms)
         self._capability = self.TIMES | self.ADJOINT_TIMES
 
     def _device_preparation(self, x, mode):
@@ -44,9 +55,16 @@ class OuterProduct(LinearOperator):
     def apply(self, x, mode):
         self._check_input(x, mode)
         self._device_preparation(x, mode)
+
+        fval = self._field.val
+        if mode == self.ADJOINT_TIMES and self._complex:
+            fval = np.conj(fval)
+        args = [fval, x.val]
+        if self._flip:
+            args = args[::-1]
         if mode == self.TIMES:
-            res = np.multiply.outer(self._field.val, x.val)
+            res = np.multiply.outer(*args)
         else:
             axes = len(self._field.shape)
-            res = np.tensordot(self._field.val, x.val, axes)
+            res = np.tensordot(*args, axes=axes)
         return Field(self._tgt(mode), res)